Fix EXTCODECOPY with empty account (#1429)
We treat empty accounts by storing their code_hash in the RwTable as 0.
EXTCODECOPY was obtaining the bytecode length by querying the bytecode
table with code_hash=0 on existing accounts, but that entry should be
invalid (there's no bytecode with code_hash=0). Skip the bytecode table
length lookup when code_hash=0.

I've also reintroduced the `Block::debug_print_txs_steps_rw_ops`
function, updated to use the new `Block::get_rws` API. This function is
not used in the code, but it's very convenient to call it when
debugging.

Resolve
https://github.com/privacy-scaling-explorations/zkevm-circuits/issues/1190

- [x] Bug fix (non-breaking change which fixes an issue)